Refactors: The user endpoints are in its own router.
