+++ index.js
@@ -1,8 +1,8 @@
 "use strict";
 
-const { response } = require('express');
 const express = require('express');
-const dh = require('./data-handler.js');
+const dh = require('./data-handler');
+const usersRouter = require('./users');
 
 const app = express();
 
@@ -73,58 +73,11 @@ function middlewareAuthorization(req, res, next){
     next();
 }
 
+app.use('/api/users', middlewareAPI);
 app.use('/api/users/:id', middlewareAPI);
 app.use('/api/users/:id', middlewareAuthentication);
 app.use('/api/users/:id', middlewareAuthorization);
-
-app.use('/api/users', middlewareAPI);
-
-app.post('/api/users', (req, res) => {
-    const result = dh.validateUser(req.body);
-
-    if(result.error){
-        res.status(400).send(result.error.details[0].message);
-        return;
-    }
-
-    const user = dh.postUser(req.body);
-
-    if (!user){
-        res.status(400).send('Email already in use for another account.');
-        return;
-    }
-
-    res.status(201).send(user);
-});
-
-app.use('/api/users', middlewareAPI);
-
-app.get('/api/users/:id', (req, res) => {
-    const user = dh.getUser(req.params.id);
-    if(!user) res.status(404).send('The user with this id does not exist.');
-    
-    res.send(user);
-});
-
-app.put('/api/users/:id', (req, res) => {
-    let user = dh.getUser(req.params.id);
-    if(!user) res.status(404).send('The user with this id does not exist.');
-
-    const result = dh.validateUser(req.body);
-    if(result.error) res.status(400).send(result.error.details[0].message);
-
-    user = dh.putUser(req.params.id, req.body);
-    res.send(user);
-});
-
-app.delete('/api/users/:id', (req, res) => {
-    let user = dh.getUser(req.params.id);
-    if(!user) res.status(404).send('The user with this id does not exist.');
-
-    dh.deleteUser(req.params.id);
-    res.status(204).send();
-});
-
+app.use('/api/users', usersRouter);
 
 const port = process.env.PORT || 3000;
 app.listen(port, () => console.log(`App listening on port ${port}!`));
